+++ examples/llm/api_server/tensorrt_llm/__init__.py
@@ -12,6 +12,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .api_server import vLLMApiServer
+from .api_server import TensorRTLLMApiServer
 
-__all__ = ["vLLMApiServer"]
+__all__ = ["TensorRTLLMApiServer"]
