+++ examples/llm/graphs/tensorrt_llm/agg.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from components.frontend import Frontend
-from components.processor import Processor
-from components.worker import TensorRTLLMWorker
+from api_server.tensorrt_llm import TensorRTLLMApiServer
+from processor.tensorrt_llm import Processor
+from worker.tensorrt_llm import TensorRTLLMWorker
 
-Frontend.link(Processor).link(TensorRTLLMWorker)
+TensorRTLLMApiServer.link(Processor).link(TensorRTLLMWorker)
