+++ examples/llm/graphs/tensorrt_llm/agg_router.py
@@ -13,12 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from components.frontend import Frontend
-from components.kv_router import Router
-from components.prefill_worker import TensorRTLLMPrefillWorker
-from components.processor import Processor
-from components.worker import TensorRTLLMWorker
+from api_server.tensorrt_llm import TensorRTLLMApiServer
+from processor.tensorrt_llm import Processor
+from router.tensorrt_llm import Router
+from worker.tensorrt_llm import TensorRTLLMWorker
 
-Frontend.link(Processor).link(Router).link(TensorRTLLMWorker).link(
-    TensorRTLLMPrefillWorker
-)
+TensorRTLLMApiServer.link(Processor).link(Router).link(TensorRTLLMWorker)
