+++ examples/llm/graphs/tensorrt_llm/configs/agg.yaml
@@ -13,17 +13,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-Frontend:
+TensorRTLLMApiServer:
   served_model_name: deepseek-ai/DeepSeek-R1-Distill-Llama-8B
   endpoint: dynamo.Processor.chat/completions
   port: 8000
 
 Processor:
-  engine_args: "configs/llm_api_config.yaml"
+  engine_args: "graphs/tensorrt_llm/configs/llm_api_config.yaml"
   router: round-robin
 
 TensorRTLLMWorker:
-  engine_args: "configs/llm_api_config.yaml"
+  engine_args: "graphs/tensorrt_llm/configs/llm_api_config.yaml"
   router: round-robin
   ServiceArgs:
     workers: 1
