+++ examples/llm/graphs/tensorrt_llm/configs/disagg.yaml
@@ -13,22 +13,31 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-Frontend:
+TensorRTLLMApiServer:
   served_model_name: deepseek-ai/DeepSeek-R1-Distill-Llama-8B
   endpoint: dynamo.Processor.chat/completions
   port: 8000
 
 Processor:
-  engine_args: "configs/llm_api_config_router.yaml"
-  router: kv
-
-Router:
-  model-name: deepseek-ai/DeepSeek-R1-Distill-Llama-8B
-  min-workers: 1
+  engine_args: "graphs/tensorrt_llm/configs/llm_api_config.yaml"
+  router: round-robin
+  remote-prefill: true
 
 TensorRTLLMWorker:
-  engine_args: "configs/llm_api_config_router.yaml"
-  router: kv
+  engine_args: "graphs/tensorrt_llm/configs/llm_api_config.yaml"
+  llmapi-disaggregated-config: "graphs/tensorrt_llm/configs/llmapi_disagg_configs/single_node_config.yaml"
+  remote-prefill: true
+  min-prefill-workers: 1
+  router: round-robin
+  ServiceArgs:
+    workers: 1
+    resources:
+      gpu: 1
+
+TensorRTLLMPrefillWorker:
+  engine_args: "graphs/tensorrt_llm/configs/llm_api_config.yaml"
+  llmapi-disaggregated-config: "graphs/tensorrt_llm/configs/llmapi_disagg_configs/single_node_config.yaml"
+  router: round-robin
   ServiceArgs:
     workers: 1
     resources:
