+++ examples/llm/graphs/tensorrt_llm/disagg_router.py
@@ -13,8 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from components.frontend import Frontend
-from components.processor import Processor
-from components.worker import TensorRTLLMWorker
+from api_server.tensorrt_llm import TensorRTLLMApiServer
+from processor.tensorrt_llm import Processor
+from router.tensorrt_llm import Router
+from worker.tensorrt_llm import TensorRTLLMPrefillWorker, TensorRTLLMWorker
 
-Frontend.link(Processor).link(TensorRTLLMWorker)
+TensorRTLLMApiServer.link(Processor).link(Router).link(TensorRTLLMWorker).link(
+    TensorRTLLMPrefillWorker
+)
