+++ examples/llm/processor/tensorrt_llm/__init__.py
@@ -12,3 +12,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .processor import Processor
+
+__all__ = ["Processor"]
